+++ src/main/java/com/dotweblabs/gwt/client/Parse.java
@@ -21,8 +21,11 @@ import com.dotweblabs.shape.client.HttpRequestException;
 import com.dotweblabs.shape.client.Shape;
 import com.google.gwt.json.client.JSONObject;
 import com.google.gwt.json.client.JSONString;
+import com.google.gwt.json.client.JSONValue;
 import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
+
+import java.util.Iterator;
 
 /**
  *
@@ -81,6 +84,37 @@ public class Parse {
             JSONObject opponents = new JSONObject();
             opponents.put("__op", new JSONString("Delete"));
             payload.put("opponents", opponents);
+            Shape.put(path)
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .body(payload.toString())
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            HttpRequestException ex = (HttpRequestException) throwable;
+                            callback.onFailure(ex);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            callback.onSuccess(ParseResponse.parse(s));
+                        }
+                    });
+        }
+        public static void update(final ParseObject ref, final AsyncCallback<ParseResponse> callback) {
+            String objectId = ref.getObjectId();
+            final String className = ref.getClassName();
+            final String path = Parse.SERVER_URL + Parse.CLASSES_URI + className + "/" + objectId;
+            JSONObject payload = new JSONObject();
+            Iterator<String> it = ref.keySet().iterator();
+            while(it.hasNext()) {
+                String key = it.next();
+                JSONValue value = ref.get(key);
+                // Do not copy objectId
+                if(!key.equals("objectId")){
+                    payload.put(key, value);
+                }
+            }
             Shape.put(path)
                     .header("X-Parse-Application-Id", X_Parse_Application_Id)
                     .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)