+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0-SNAPSHOT - 06/29/16
 
 ### Added
+- Update ParseObject function
 - Delete ParseObject by objectId
 - Create ParseObject and save to parse-server
 - Initial commit